# sdk/go/api/openapi.yaml
openapi: 3.0.3
info:
  description: "REST API for Meept daemon - exposes chat, memory, task queue, skills,\
    \ and self-improvement capabilities"
  title: Meept HTTP API
  version: 0.2.0
servers:
- description: Local development server
  url: http://localhost:8081
...
paths:
...
  /api/v1/chat:
    post:
      responses:
        "200":
          description: Success
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: s.handleChat
      tags:
      - v1
...
  /api/v1/chat/stream:
    get:
      responses:
        "200":
          description: Success
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: s.handleChatStream
      tags:
      - v1
...
  /api/v1/memory/query:
    post:
      responses:
        "200":
          description: Success
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: s.handleMemoryQuery
      tags:
      - v1
...
  /api/v1/sessions:
    get:
      responses:
        "200":
          description: Success
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: s.handleSessionList
      tags:
      - v1
  /api/v1/sessions/{id}:
    delete:
      responses:
        "200":
          description: Success
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: s.handleSessionDelete
      tags:
      - v1
...
  /api/v1/skills:
    get:
      responses:
        "200":
          description: Success
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: s.handleSkillsList
      tags:
      - v1
...
  /api/v1/skills/{slug}/execute:
    post:
      responses:
        "200":
          description: Success
        "400":
          description: Bad request
        "401":
          description: Unauthorized
        "500":
          description: Internal server error
      security:
      - ApiKeyAuth: []
      summary: s.handleSkillsExecute
      tags:
      - v1
...
  /ws:
    get:
      description: |
        Establishes a WebSocket connection for receiving real-time agent progress events.
        Clients can subscribe to specific sessions via subscribe messages.
      operationId: getWebSocket
      parameters:
      - description: Optional session ID to filter events
        explode: true
        in: query
        name: session_id
        required: false
        schema:
          type: string
        style: form
      responses:
        "101":
          description: Switching Protocols - WebSocket established
      summary: WebSocket connection for real-time updates
      tags:
      - WebSocket
components:
...
  securitySchemes:
    ApiKeyAuth:
      description: "API key authentication. Use: Authorization: Bearer YOUR_API_KEY"
      in: header
      name: Authorization
      type: apiKey